clean up modulemanager, it doesn't hold the systems anymore.

+++ src/ManaForgeEngine/Public/ModelData.h
@@ -7,7 +7,7 @@ class ModelData
 public:
 	virtual ~ModelData() = default;
 
-	virtual void init(const vec2* vertLocs, const vec2* UVs, uint32 numVerts, const uvec3* elems, uint32 numElems) = 0;
+	virtual void init(const vec2* vertLocs, const vec2* UVs, size_t numVerts, const uvec3* elems, size_t numElems) = 0;
 	virtual bool isInitialized() = 0;
 	// TODO:
 	// virtual void init(const path_t& path)
